created base service for open ai api
added new functionality for open ai api, send requests

+++ internal/model/open-ai.model.go
@@ -4,7 +4,7 @@ type OpenAIRequest struct {
 	Model       string    `json:"model"`
 	Messages    []Message `json:"messages"`
 	MaxTokens   int       `json:"max_tokens"`
-	Temperature float64   `json:"temperature"`
+	Temperature float32   `json:"temperature"`
 }
 
 type Message struct {
